requirement.txt
absl-py==0.11.0
appdirs==1.4.4
argon2-cffi==20.1.0
astor==0.8.1
astroid==2.4.2
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
autopep8==1.5.4
backcall==0.2.0
bleach==3.2.1
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
click==8.0.1
cloudpickle==1.6.0
cmake==3.18.4.post1
colorama==0.4.4
cudnn-python-wrappers==1.0
cycler==0.10.0
Cython==0.29.22
decorator==4.4.2
deepface==0.0.65
defusedxml==0.6.0
distlib==0.3.1
dlib @ file:///C:/Users/001si/Desktop/face_recog_dlib_file/dlib-19.19.0-cp37-cp37m-win_amd64.whl
entrypoints==0.3
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.0.12
Flask==2.0.1
Flask-Cors==3.0.10
future==0.18.2
gast==0.3.3
gdown==3.13.0
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
h5py==2.10.0
hyperopt==0.2.5
idna==2.10
imageio==2.9.0
importlib-metadata==2.0.0
imutils==0.5.3
intel-openmp==2021.1.2
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
isort==5.6.4
itsdangerous==2.0.1
jedi==0.17.2
Jinja2==3.0.1
joblib==0.17.0
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.7.0
jupyterlab==2.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
jupyterthemes==0.20.0
kaggle==1.5.12
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
lesscpy==0.14.0
Markdown==3.3.3
MarkupSafe==2.0.1
matplotlib==3.3.3
mccabe==0.6.1
mistune==0.8.4
mkl==2021.1.1
mtcnn==0.1.1
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
networkx==2.5
notebook==6.1.5
numpy==1.18.5
oauthlib==3.1.0
opencv-contrib-python==4.4.0.46
opencv-python==4.4.0.46
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.4
pandocfilters==1.4.3
parso==0.7.1
pickleshare==0.7.5
Pillow==8.0.1
ply==3.11
prometheus-client==0.9.0
prompt-toolkit==3.0.8
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.7.2
pylint==2.6.0
pymongo==3.11.1
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.1
python-slugify==5.0.2
pytz==2020.4
PyWavelets==1.1.1
pywin32==300
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==20.0.0
requests==2.24.0
requests-oauthlib==1.3.0
retina-face==0.0.4
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.23.2
scikit-plot==0.3.7
scipy==1.4.1
seaborn==0.11.1
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
tbb==2021.1.1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
tensorflow-gpu==2.3.1
tensorflow-gpu-estimator==2.3.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
text-unidecode==1.3
tflearn==0.5.0
threadpoolctl==2.1.0
tifffile==2020.11.18
toml==0.10.2
torch==1.8.1
torchvision==0.9.1
tornado==6.1
tqdm==4.53.0
traitlets==5.0.5
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.11
virtualenv==20.4.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.4.0
